consolidate apps service
