+++ api-node/src/app.module.ts
@@ -3,7 +3,6 @@ import { PackagesController } from './packages/packages.controller';
 import { HealthCheckController } from './health/healthCheck.controller';
 import { MarketingController } from './marketing/marketing.controller';
 import { AppsController } from './apps/apps.controller';
-import { AppsService } from './apps/apps.service';
 import { SdksController } from './sdks/sdks.controller';
 import { AwsLambdaLayersController } from './aws-lambda-layers/aws-lambda-layers.controller';
 import { RegistryService } from './registry/registry.service';
@@ -18,6 +17,6 @@ import { RegistryService } from './registry/registry.service';
     SdksController,
     AwsLambdaLayersController,
   ],
-  providers: [AppsService, RegistryService],
+  providers: [RegistryService],
 })
 export class AppModule {}
