--- api-node/src/apps/apps.service.ts
@@ -1,94 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import * as fs from 'fs';
-import * as path from 'path';
-import { AppEntry, AppsResponse } from './types';
-
-const APPS_PATH = path.join('..', 'apps');
-
-@Injectable()
-export class AppsService {
-  getApps(): AppsResponse {
-    try {
-      const apps = fs.readdirSync(APPS_PATH);
-      return apps.reduce((acc, appName) => {
-        try {
-          const app = this.getApp(appName);
-          if (app) {
-            acc[appName] = app;
-          }
-        } catch {
-          // Continue to next iteration if there's an error
-        }
-        return acc;
-      }, {});
-    } catch (error) {
-      // Handle error (e.g., log it or throw a custom exception)
-      console.error('Error reading apps directory:', error);
-    }
-  }
-
-  getApp(appId: string, version: string = 'latest'): AppEntry | null {
-    try {
-      const filePath = path.join(APPS_PATH, appId, `${version}.json`);
-      const fileContent = fs.readFileSync(filePath, 'utf8');
-      return JSON.parse(fileContent);
-    } catch {
-      return null;
-    }
-  }
-
-  findDownloadUrl(
-    appInfo: AppEntry,
-    pkgName: string,
-    arch: string,
-    platform: string,
-  ): string | null {
-    const normalizedPackage = pkgName
-      .replace(/_/g, '-')
-      .toLowerCase()
-      .split('-');
-    for (const [_, url] of Object.entries(appInfo.file_urls)) {
-      let normalizedUrl = url.toLowerCase();
-      if (normalizedUrl.endsWith('.exe')) {
-        normalizedUrl = normalizedUrl.slice(0, -4);
-      }
-      const parts = normalizedUrl.split('/').pop()!.split('-');
-      if (
-        parts.length > 2 &&
-        parts.slice(0, -2).join('-') === normalizedPackage.join('-') &&
-        parts[parts.length - 1].replace(/_/g, '-') ===
-          arch.toLowerCase().replace(/_/g, '-') &&
-        parts[parts.length - 2] === platform.toLowerCase()
-      ) {
-        return url;
-      }
-    }
-    return null;
-  }
-
-  getUrlChecksums(
-    appInfo: AppEntry,
-    url: string,
-  ): Record<string, string> | null {
-    for (const fileInfo of Object.values(appInfo.files)) {
-      if (fileInfo.url === url) {
-        return (fileInfo as any).checksums || null;
-      }
-    }
-    return null;
-  }
-
-  makeDigest(checksums: Record<string, string> | null): string {
-    if (!checksums) return '';
-    const digestParts: string[] = [];
-    for (const [algo, value] of Object.entries(checksums)) {
-      if (algo.endsWith('-hex')) {
-        const base64Value = Buffer.from(value, 'hex').toString('base64');
-        digestParts.push(`${algo.slice(0, -4)}=${base64Value}`);
-      } else if (algo.endsWith('-base64')) {
-        digestParts.push(`${algo.slice(0, -7)}=${value}`);
-      }
-    }
-    return digestParts.join(',');
-  }
-}
