+++ api-node/src/aws-lambda-layers/aws-lambda-layers.controller.ts
@@ -3,6 +3,8 @@ import { RegistryService } from '../registry/registry.service';
 
 @Controller('aws-lambda-layers')
 export class AwsLambdaLayersController {
+  // TODO: types
+
   constructor(private readonly registryService: RegistryService) {}
 
   @Get()
